requirements.txt
Flask
Redis
gunicorn
mysql-connector-python
watchdog